refactor(terminal): split shellRegistry.ts god-file (DEBT #4)
Split src/terminal/shellRegistry.ts (686 LOC) into cohesive siblings,
keeping the registry map + lifecycle facade in shellRegistry.ts (473):

- shellAppearance.ts (40): ShellAppearance type, setShellAppearanceProvider,
  readAppearance — the DEBT #6 appearance-provider indirection, preserved
  verbatim (no static useStore import).
- shellInput.ts (151): installInputListeners (IME/paste/selection/key).
- shellTypes.ts (34): the Shell type, shared to avoid a registry<->input cycle.

shellRegistry.ts re-exports Shell/ShellAppearance/setShellAppearanceProvider
so every importer is unchanged. Verbatim move, no runtime change.
madge --circular clean (no terminal->store cycle reintroduced).
