+++ src/terminal/shellRegistry.ts
@@ -23,79 +23,14 @@ import { WebLinksAddon } from '@xterm/addon-web-links';
 import { ClipboardAddon } from '@xterm/addon-clipboard';
 import { Unicode11Addon } from '@xterm/addon-unicode11';
 import { CanvasAddon } from '@xterm/addon-canvas';
-import { SCROLLBACK_LINES_DEFAULT, TERMINAL_FONT_SIZE_DEFAULT } from '../stores/slices/types';
 import { warn, log } from '../shared/log';
-import { pasteIntoActivePty } from './paste';
+import { readAppearance } from './shellAppearance';
+import { installInputListeners } from './shellInput';
+import type { Shell } from './shellTypes';
 
-// Appearance defaults (scrollback cap + terminal font size) are owned by
-// the store, but a static `import { useStore }` here would close a module
-// cycle: store.ts → sessionCrudSlice → terminal/shellRegistry → store.ts
-// (DEBT #6). Instead the store registers a lazy provider at boot via
-// `setShellAppearanceProvider`. Until that runs (tests / non-renderer
-// contexts) we fall back to the compile-time defaults from `slices/types`,
-// which is the same value the store's appearance slice initialises with.
-export type ShellAppearance = {
-  scrollbackLines: number;
-  terminalFontSizePx: number;
-};
-
-let appearanceProvider: (() => ShellAppearance) | null = null;
-
-/**
- * Register the live appearance source. Called once by `stores/store.ts`
- * after the store is created so `createShell` can read the user's current
- * scrollback / font-size without importing the store (which would form a
- * cycle). Idempotent re-registration is allowed (last writer wins).
- */
-export function setShellAppearanceProvider(provider: () => ShellAppearance): void {
-  appearanceProvider = provider;
-}
-
-function readAppearance(): ShellAppearance {
-  if (appearanceProvider) {
-    try {
-      return appearanceProvider();
-    } catch (e) {
-      warn('shell', 'appearance provider threw — using defaults', e);
-    }
-  }
-  return {
-    scrollbackLines: SCROLLBACK_LINES_DEFAULT,
-    terminalFontSizePx: TERMINAL_FONT_SIZE_DEFAULT,
-  };
-}
-
-export type Shell = {
-  sid: string;
-  /** DOM wrapper parented under the host. Owns the xterm canvases AND the
-   *  per-shell mask div. Never reparented after first attach — visibility
-   *  is toggled via `display` + `z-index` only. */
-  wrapper: HTMLDivElement;
-  /** Mask div sitting inside the wrapper. Shown during cold-start /
-   *  reload-of-top, hidden otherwise. */
-  mask: HTMLDivElement;
-  term: Terminal;
-  fit: FitAddon;
-  /** `pty.onData` unsubscriber. Lives for the lifetime of the shell. */
-  dataUnsubscribe: () => void;
-  /** Disposers for textarea-bound listeners (IME, paste, selection-copy). */
-  inputDisposers: Array<() => void>;
-  /** Composition state — `pty.onData` writes are buffered while true so the
-   *  IME preview box doesn't jump. */
-  composing: boolean;
-  composingBuffer: string[];
-  /** Set true after the first cold-start completes; used by the host to
-   *  know whether a re-show should trigger any mask UI (it should not). */
-  warmed: boolean;
-  /** Set true once `term.onData → pty.input` has been wired for this shell.
-   *  Used by `usePtyAttachShell` to avoid double-wiring across reload — the
-   *  Terminal instance is kept across reloads so the existing disposable
-   *  is still live. */
-  inputWired: boolean;
-  /** Pending font size to apply on next show — set when a Ctrl+wheel zoom
-   *  happens while this shell is hidden. */
-  pendingFontSize: number | null;
-};
+export type { Shell } from './shellTypes';
+export type { ShellAppearance } from './shellAppearance';
+export { setShellAppearanceProvider } from './shellAppearance';
 
 const shells: Map<string, Shell> = new Map();
 let topSid: string | null = null;
@@ -488,154 +423,6 @@ export function applyTerminalScrollback(n: number): void {
     } catch (e) {
       warn('shell', 'apply scrollback failed', e);
     }
-  }
-}
-
-// Input-side listeners: IME composition guard, capture-phase paste,
-// selection-to-clipboard auto-copy, Ctrl/Cmd+C/V/A handler. Same surface
-// as the legacy registry (these are UX features, not race handling).
-function installInputListeners(shell: Shell): void {
-  const { term, sid, wrapper } = shell;
-  const ta = term.textarea;
-  if (ta) {
-    const onStart = () => {
-      shell.composing = true;
-      shell.composingBuffer.length = 0;
-    };
-    const onEnd = () => {
-      shell.composing = false;
-      if (shell.composingBuffer.length > 0) {
-        const pending = shell.composingBuffer.join('');
-        shell.composingBuffer.length = 0;
-        try {
-          term.write(pending);
-        } catch {
-          /* best-effort */
-        }
-      }
-    };
-    ta.addEventListener('compositionstart', onStart);
-    ta.addEventListener('compositionend', onEnd);
-    shell.inputDisposers.push(() => {
-      try {
-        ta.removeEventListener('compositionstart', onStart);
-      } catch {
-        /* ignore */
-      }
-      try {
-        ta.removeEventListener('compositionend', onEnd);
-      } catch {
-        /* ignore */
-      }
-    });
-  }
-
-  let keyboardPasteHandled = false;
-  const onPasteCapture = (e: ClipboardEvent): void => {
-    e.stopImmediatePropagation();
-    e.preventDefault();
-    if (keyboardPasteHandled) {
-      keyboardPasteHandled = false;
-      return;
-    }
-    const text = e.clipboardData?.getData('text/plain') ?? '';
-    void pasteIntoActivePty(() => shell.term, sid, text || undefined);
-  };
-  wrapper.addEventListener('paste', onPasteCapture, true);
-  shell.inputDisposers.push(() => {
-    try {
-      wrapper.removeEventListener('paste', onPasteCapture, true);
-    } catch {
-      /* ignore */
-    }
-  });
-
-  try {
-    const selDisposable = term.onSelectionChange(() => {
-      const sel = term.getSelection();
-      if (sel) {
-        try {
-          window.ccsmPty?.clipboard?.writeText(sel);
-        } catch {
-          /* ignore */
-        }
-      }
-    });
-    shell.inputDisposers.push(() => {
-      try {
-        selDisposable?.dispose?.();
-      } catch {
-        /* ignore */
-      }
-    });
-  } catch (e) {
-    warn('shell', 'onSelectionChange attach failed', e);
-  }
-
-  const pasteFromClipboard = (): void => {
-    keyboardPasteHandled = true;
-    setTimeout(() => {
-      keyboardPasteHandled = false;
-    }, 0);
-    let text: string | undefined;
-    try {
-      text = window.ccsmPty?.clipboard?.readText() || undefined;
-    } catch {
-      /* best-effort */
-    }
-    void pasteIntoActivePty(() => shell.term, sid, text);
-  };
-  try {
-    term.attachCustomKeyEventHandler((ev) => {
-      if (ev.type !== 'keydown') return true;
-      const mod = ev.ctrlKey || ev.metaKey;
-      if (!mod || ev.altKey) return true;
-      const isC = ev.key === 'C' || ev.key === 'c';
-      const isV = ev.key === 'V' || ev.key === 'v';
-      const isA = ev.key === 'A' || ev.key === 'a';
-      if (!ev.shiftKey && isA) {
-        try {
-          term.selectAll();
-        } catch {
-          /* ignore */
-        }
-        return false;
-      }
-      if (!ev.shiftKey && isC) {
-        const sel = term.getSelection();
-        if (sel) {
-          try {
-            window.ccsmPty?.clipboard?.writeText(sel);
-          } catch {
-            /* ignore */
-          }
-          return false;
-        }
-        return true;
-      }
-      if (!ev.shiftKey && isV) {
-        pasteFromClipboard();
-        return false;
-      }
-      if (ev.shiftKey && isC) {
-        const sel = term.getSelection();
-        if (sel) {
-          try {
-            window.ccsmPty?.clipboard?.writeText(sel);
-          } catch {
-            /* ignore */
-          }
-        }
-        return false;
-      }
-      if (ev.shiftKey && isV) {
-        pasteFromClipboard();
-        return false;
-      }
-      return true;
-    });
-  } catch (e) {
-    warn('shell', 'attachCustomKeyEventHandler failed', e);
   }
 }
